backend/pyproject.toml
[tool.poetry]
name = "user-system"
version = "0.1.0"
description = ""
authors = ["weijay <weijay0804@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
pymysql = "^1.1.1"
cryptography = "^42.0.8"
passlib = "^1.7.4"
bcrypt = "4.0.1"
fastapi-mail = "^1.4.1"
pyjwt = "^2.8.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
